Bajaj/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
PyMuPDF==1.24.7
python-docx==1.1.2
numpy==1.26.4
faiss-cpu==1.7.4
tenacity==8.3.0
orjson==3.10.6
